% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_optim_MM.R
\name{bayesian_optimization_CCA}
\alias{bayesian_optimization_CCA}
\title{Bayesian Hyperparameter optimization
Get optimal penalties using bayesian optimization.
Score function depends on whether sparsity is assumed or not}
\usage{
bayesian_optimization_CCA(
  C,
  p,
  q,
  n,
  alpha_x,
  alpha_y,
  order,
  low_a,
  low_b,
  tol,
  lr,
  epochs,
  lr_decay,
  criterion = "TPO",
  init_ortho = TRUE
)
}
\arguments{
\item{C}{estimated covariance matrix}

\item{p}{number of variables in data_x}

\item{q}{number of variables in data_y}

\item{n}{number of observations}

\item{alpha_x}{(1xk) A positive real vector containing numbers between 0 and
1, indicating the elastic
 net parameter for data_x for association of order k.}

\item{alpha_y}{(1xk) A positive real vector containing numbers between 0 and
1, indicating the elastic
 net parameter for data_y for association of order k.}

\item{order}{The order of association to be computed}

\item{low_a}{A (px(k-1)) matrix of lower order linear combinations}

\item{low_b}{A (qx(k-1)) matrix of lower order linear combinations}

\item{tol}{tolerance for training algorithm}

\item{lr}{learning rate for gradient descent training algorithm}

\item{epochs}{maximum number of epochs for gradient descent training algorithm}

\item{lr_decay}{parameter for}

\item{criterion}{type of criterion used to determine optimal hyperparameters,
default: criterion = TPO}

\item{init_ortho}{logical}
}
\value{
a list object containing the optimal hyperparameters and a summary
of the optimization process
}
\description{
Bayesian Hyperparameter optimization
Get optimal penalties using bayesian optimization.
Score function depends on whether sparsity is assumed or not
}
