% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_optim_MM.R
\name{bayesian_optimization_PCA_SVD}
\alias{bayesian_optimization_PCA_SVD}
\title{Bayesian Hyperparameter optimization for PCA via SVD}
\usage{
bayesian_optimization_PCA_SVD(
  X,
  groups,
  rsvd,
  n,
  p,
  k,
  alpha_x,
  loss_type,
  h,
  tol,
  lr,
  epochs,
  lr_decay,
  bounds_input,
  criterion = "TPO"
)
}
\arguments{
\item{rsvd}{object resulting from SVD on R}

\item{n}{number of observations}

\item{p}{number of variables in data_x}

\item{alpha_x}{scalar: A positive real scalar between 0 and
1, indicating the elastic
 net parameter for data_x.}

\item{tol}{tolerance for training algorithm}

\item{lr}{learning rate for gradient descent training algorithm}

\item{epochs}{maximum number of epochs for gradient descent training algorithm}

\item{lr_decay}{parameter for}

\item{criterion}{type of criterion used to determine optimal hyperparameters,
default: criterion = TPO}

\item{Q}{Q from QR decomposition on data_x}

\item{R}{R from QR decomposition on data_x}
}
\value{
a list object containing the optimal hyperparameters and a summary
of the optimization process
}
\description{
Get optimal penalties using bayesian optimization.
Score function depends on whether sparsity is assumed or not
}
