% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constraint_fun}
\alias{constraint_fun}
\title{Domain constraint for robust and sparse PCA / CCA}
\usage{
constraint_fun(z, C = NULL)
}
\arguments{
\item{z}{A (1xp) torch::torch_tensor}

\item{C}{A (pxp) quadratic matrix}
}
\value{
z_constrained, the value of the equality constraints
}
\description{
Function to compute domain constraint for a tensor
}
