% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f1_score}
\alias{f1_score}
\title{F1 score}
\usage{
f1_score(ind_outl, flag_outl)
}
\arguments{
\item{ind_outl}{position of outliers}

\item{flag_outl}{flagged cells}
}
\description{
Function to compute F1 score for outlier detection, given the true positions
}
