% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{higher_order}
\alias{higher_order}
\title{Higher order constraint for robust and sparse PCA / CCA}
\usage{
higher_order(z, low_z, C = NULL)
}
\arguments{
\item{z}{A (1xp) torch::torch_tensor}

\item{low_z}{A (kxp) torch::torch_tensor}

\item{C}{A (pxp) matrix}
}
\description{
Function to compute higher order constraint for a tensor
}
