% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loss_huber}
\alias{loss_huber}
\title{Loss function for robust and sparse PCA via SVD}
\usage{
loss_huber(M, d = 1.35)
}
\arguments{
\item{M}{A matrix torch::torch_tensor}

\item{d}{Scalar, parameter for Huber loss}
}
\value{
A 1x1 torch::torch_tensor
}
\description{
Function to compute Pseudo-Huber loss for a given matrix.
}
