% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loss_lts}
\alias{loss_lts}
\title{Loss function for robust and sparse PCA via SVD}
\usage{
loss_lts(M, h = 0.75)
}
\arguments{
\item{M}{A matrix torch::torch_tensor}

\item{h}{Trimming parameter (robustness) of observations to use per column}
}
\value{
A 1x1 torch::torch_tensor
}
\description{
Function to compute trimmes loss for a given matrix.
}
