% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loss_tukey}
\alias{loss_tukey}
\title{Loss function for robust and sparse PCA via SVD}
\usage{
loss_tukey(M, k = 1.35)
}
\arguments{
\item{M}{A matrix torch::torch_tensor}

\item{k}{Scalar, parameter for Huber loss}
}
\value{
A 1x1 torch::torch_tensor
}
\description{
Function to compute Tukey loss for a given matrix.
}
