% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_CCA_MM.R
\name{model_CCA}
\alias{model_CCA}
\title{NN module for robust and sparse CCA}
\usage{
model_CCA(
  C,
  p,
  q,
  pen_x,
  pen_y,
  order,
  alpha_x,
  alpha_y,
  rho,
  warm_a = NA,
  warm_b = NA,
  warm_u = NA,
  low_a = NA,
  low_b = NA,
  init_ortho = TRUE
)
}
\arguments{
\item{C}{estimated covariance matrix}

\item{p}{number of variables in data_x}

\item{q}{number of variables in data_y}

\item{pen_x}{(1xk) A positive real vector containing the penalty parameter for
data_x for association of order k.}

\item{pen_y}{(1xk) A positive real vector containing the penalty parameter for
data_y for association of order k.}

\item{order}{The order of association to be computed}

\item{alpha_x}{(1xk) A positive real vector containing numbers between 0 and
1, indicating the elastic
 net parameter for data_x for association of order k.}

\item{alpha_y}{(1xk) A positive real vector containing numbers between 0 and
1, indicating the elastic
 net parameter for data_y for association of order k.}

\item{rho}{A real number. The current value of the penalty strength in the
Augmented Lagrangian function.}

\item{warm_a}{A (px1) vector of the currently best linear combination}

\item{warm_b}{A (qx1) vector of the currently best linear combination}

\item{warm_u}{A vector of the currently best multiplier values}

\item{low_a}{A (px(k-1)) matrix of lower order linear combinations}

\item{low_b}{A (qx(k-1)) matrix of lower order linear combinations}

\item{init_ortho}{logical: whether initialization should be orthogonal
for higher-order directions}
}
\description{
NN module for robust and sparse CCA
}
