% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_PCA_SVD.R
\name{model_PCA_SVD}
\alias{model_PCA_SVD}
\title{NN module for robust and sparse PCA via SVD}
\usage{
model_PCA_SVD(
  X,
  groups,
  rsvd,
  n,
  p,
  pen_x,
  alpha_x,
  loss_type,
  h,
  warm_v = NA,
  warm_m = NA
)
}
\arguments{
\item{X}{data input}

\item{rsvd}{object containing SVD decompositin of R}

\item{n}{number of}

\item{p}{number of variables in data_x}

\item{pen_x}{(1xk) A positive real vector containing the penalty parameter for
data_x for association of order k.}

\item{alpha_x}{(1xk) A positive real vector containing numbers between 0 and
1, indicating the elastic
 net parameter for data_x for association of order k.}

\item{warm_v}{A vector of the currently best linear combination}

\item{warm_m}{A vector of the currently best multiplier values}

\item{order}{The order of association to be computed}

\item{warm_u}{A vector of the currently best linear combination}

\item{warm_s}{A  scalar of the currently best singular value}

\item{low_u}{A (px(k-1)) matrix of lower order linear combinations}
}
\description{
NN module for robust and sparse PCA via SVD
}
