% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CCA_MM.R
\name{plot.CCA_result}
\alias{plot.CCA_result}
\title{Plot CCA
Function to plot summary of CCA computation and hyperparameter optimization}
\usage{
\method{plot}{CCA_result}(x, ...)
}
\arguments{
\item{x}{object of class CCA_result}

\item{...}{additional parameters to be passed to ggplot2}
}
\description{
Plot CCA
Function to plot summary of CCA computation and hyperparameter optimization
}
