% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_PCA_SVD.R
\name{plot.PCA_result}
\alias{plot.PCA_result}
\title{Plot PCA
Function to plot summary of PCA computation and hyperparameter optimization}
\usage{
\method{plot}{PCA_result}(x, ...)
}
\arguments{
\item{x}{object of class PCA_result}

\item{...}{additional parameters to be passed to ggplot2}
}
\description{
Plot PCA
Function to plot summary of PCA computation and hyperparameter optimization
}
