% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{principal_angles}
\alias{principal_angles}
\title{Get principal angles.}
\usage{
principal_angles(a, b)
}
\arguments{
\item{a}{A matrix.}

\item{b}{A matrix.}
}
\value{
A list of angles.
}
\description{
Calculate minimal angle between subspace a and b
}
