% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_spearman.R
\docType{data}
\name{res_spearman}
\alias{res_spearman}
\title{Results of ccaMM() demo}
\format{
An object of class CCA_result
}
\source{
The object can be created by running the code in the "demo_CCA" vignette.
}
\usage{
data(res_spearman)
}
\description{
A dataset containing results for speeding up the build of the "demo_CCA" vignette.
}
\examples{
# data(res_spearman)

}
\keyword{datasets}
