% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TNR}
\alias{TNR}
\title{True negative rate}
\usage{
TNR(A_true, A)
}
\arguments{
\item{A_true}{A matrix of the ground truth.}

\item{A}{A matrix.}
}
\value{
The number of correctly identified zero components.
}
\description{
This function compares the structure of two matrices A and A_true.
}
