% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TPR}
\alias{TPR}
\title{True positive rate}
\usage{
TPR(A_true, A)
}
\arguments{
\item{A_true}{A matrix of the ground truth.}

\item{A}{A matrix.}
}
\value{
The number of correctly identified non-zero components.
}
\description{
This function compares the structure of two matrices A and A_true.
}
