% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_PCA_SVD.R
\name{train_PCA_SVD}
\alias{train_PCA_SVD}
\title{MM algorithm for PCA objects
implements MM algorithms for robust and sparse PCA}
\usage{
train_PCA_SVD(
  X,
  groups,
  rsvd,
  n,
  p,
  pen_x,
  alpha_x,
  loss_type,
  h,
  tol,
  lr,
  epochs,
  lr_decay,
  show_warnings = FALSE,
  init_ortho = TRUE
)
}
\arguments{
\item{p}{number of variables in data_x}

\item{pen_x}{(1xk) A positive real vector containing the penalty parameter for
data_x for association of order k.}

\item{alpha_x}{(1xk) A positive real vector containing numbers between 0 and
1, indicating the elastic
 net parameter for data_x for association of order k.}

\item{tol}{tolerance for training algorithm}

\item{lr}{learning rate for gradient descent training algorithm}

\item{epochs}{maximum number of epochs for gradient descent training algorithm}

\item{lr_decay}{parameter for}

\item{show_warnings}{logical}

\item{init_ortho}{logical}

\item{C}{estimated covariance matrix}

\item{low_a}{A (px(k-1)) matrix of lower order linear combinations}
}
\description{
MM algorithm for PCA objects
implements MM algorithms for robust and sparse PCA
}
